✨ feat: Amr 별 사용률 worst 3 조회 추가

+++ be/smartfactory/src/main/java/info/smartfactory/domain/dashboard/service/AmrPercentDto.java
@@ -8,5 +8,5 @@ import lombok.*;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class AmrPercentDto {
     private Long amrId;
-    private int percentage;
+    private Long percentage;
 }
